+++ R/sandbox.R
@@ -137,10 +137,9 @@ renv_sandbox_generate <- function(sandbox) {
   renv_scope_lock(lockfile)
 
   # make the library temporarily writable
-  lock <- getOption("renv.sandbox.locking_enabled") %||% all(
-    !renv_package_checking() &&
-    !renv_path_within(sandbox, tempdir())
-  )
+  lock <- getOption("renv.sandbox.locking_enabled") %||% {
+    !renv_package_checking() && !renv_path_within(sandbox, tempdir())
+  }
 
   if (lock) {
     dlog("sandbox", "Temporarily unlocking sandbox.")
@@ -230,7 +229,7 @@ renv_sandbox_unlock <- function(sandbox = NULL, project = NULL) {
   Sys.chmod(sandbox, mode = "0755")
 }
 
-#' The system library sandbox
+#' The default library sandbox
 #'
 #' @description
 #' An \R installation can have up to three types of library paths available
@@ -244,38 +243,38 @@ renv_sandbox_unlock <- function(sandbox = NULL, project = NULL) {
 #'   system are installed. This library path, if it exists, is visible to all
 #'   users on the system.
 #'
-#' - The _system library_, where \R packages distributed with \R itself are
+#' - The _default library_, where \R packages distributed with \R itself are
 #'   installed. This library path is visible to all users on the system.
 #'
 #' Normally, only so-called "base" and "recommended" packages should be installed
-#' in the system library (You can get a list of these packages with
+#' in the default library. (You can get a list of these packages with
 #' `installed.packages(priority = c("base", "recommended"))`). However, it is
-#' possible for users and administrators to install packages into the system
+#' possible for users and administrators to install packages into the default
 #' library, if the filesystem permissions permit them to do so. (This, for
 #' example, is the default behavior on macOS.)
 #'
-#' Because the site and system libraries are visible to all users, having those
+#' Because the site and default libraries are visible to all users, having those
 #' accessible in renv projects can potentially break isolation -- that is,
-#' if a package were updated in the system library, that update would be visible
+#' if a package were updated in the default library, that update would be visible
 #' to all \R projects on the system.
 #'
-#' To help defend against this, renv uses something it calls the "system library
-#' sandbox", or the "sandbox", to isolate renv projects from non-"base" packages
-#' that are installed into the system library. To accomplish this, when an renv
-#' project is loaded, renv will:
+#' To help defend against this, renv uses something called the "sandbox" to
+#' isolate renv projects from non-"base" packages that are installed into the
+#' default library. When an renv project is loaded, renv will:
 #'
 #' - Create a new, empty library path (called the "sandbox"),
 #'
-#' - Link only the "base" packages from the system library into the sandbox,
+#' - Link only the "base" and "recommended" packages from the default library
+#'   into the sandbox,
 #'
 #' - Mark the sandbox as read-only, so that users are unable to install packages
 #'   into this library,
 #'
-#' - Instruct the \R session to use the "sandbox" as the system library.
+#' - Instruct the \R session to use the "sandbox" as the default library.
 #'
 #' This process is mostly transparent to the user. However, because the sandbox
 #' is read-only, if you later need to remove the sandbox, you'll need to reset
-#' file permissions manually; for example, with `renv::sandbox$unlock()`
+#' file permissions manually; for example, with `renv::sandbox$unlock()`.
 #'
 #' If you'd prefer to keep the sandbox unlocked, you can also set:
 #'
@@ -290,6 +289,9 @@ renv_sandbox_unlock <- function(sandbox = NULL, project = NULL) {
 #' ```
 #' RENV_CONFIG_SANDBOX_ENABLED = FALSE
 #' ```
+#'
+#' The sandbox library path can also be configured using the `RENV_PATHS_SANDBOX`
+#' environment variable: see [paths] for more details.
 #'
 #' @format NULL
 #' @export