+++ NEWS.md
@@ -4,6 +4,9 @@
 * If `renv::snapshot()` finds missing packages, a new prompt allows you to 
   install them before continuing (#1198).
 
+* When prompting to activate a project, you now get three options that hopefully
+  make your choices more clear.
+  
 * `renv::dependencies()` no longer treats `box::use(module/file)` as using
   package `module` (#1377).